+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # ZLUX Angular File Tree
 
+## 0.10.0
+
+* Added a right click Download file feature
+
 ## 0.9.0
 
-* Added a right click Upload file feature
+* Added a right click Upload file feature, NOTE: BINARY is still unsupported
 
 ## 0.8.0